+++ src/TheChest.Inventories/Containers/Inventory.cs
@@ -216,7 +216,7 @@ namespace TheChest.Inventories.Containers
                 }
             }
             
-            return default;
+            return default!;
         }
         /// <inheritdoc/>
         /// <remarks>