+++ tests/TheChest.Inventories.Tests/Containers/StackInventory/StackInventoryTests.Replace.cs
@@ -27,6 +27,9 @@
             );
         }
 
+        //TODO: add tests to slot with more items than stack size
+        //TODO: add tests to calls and does not call OnReplace event
+
         [Test]
         public void Replace_NullItems_DoesNotReplace()
         {
@@ -66,7 +69,9 @@
             var inventory = this.containerFactory.FullContainer(size, stackSize, item);
 
             var randomIndex = this.random.Next(0, size);
-            var newItems = this.itemFactory.CreateManyRandom(this.random.Next(1, 20));
+            var newStackSize = this.random.Next(1, stackSize - 1);
+            var newItems = this.itemFactory.CreateManyRandom(newStackSize);
+
             var result = inventory.Replace(newItems, randomIndex);
 
             Assert.That(result, Has.Length.EqualTo(stackSize));