+++ tests/TheChest.Inventories.Tests/Usings.cs
@@ -3,4 +3,6 @@ global using TheChest.Inventories.Tests.Containers.Factories.Interfaces;
  
 global using TheChest.Core.Tests.Slots.Factories.Interfaces;
 global using TheChest.Inventories.Tests.Extensions;
- 
+
+global using TheChest.Inventories.Tests.Containers.Extensions;
+global using TheChest.Inventories.Tests.Slots.Extensions;